backend/requirements.txt
flask
flask-sqlalchemy
flask-marshmallow
marshmallow-sqlalchemy
flask-cors
pymysql
cryptography
python-dotenv
google-generativeai
pytesseract
pillow
werkzeug
gunicorn
